DTOs for Obstacles, Pickups, Players and Score Locations
