--- UnityProject/Assets/Scripts/DTOs/ObstaclesDTO.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-[Serializable]
-public struct ObstaclesDTO
-{
-    public ObstacleDTO[] obstacles;
-}
