--- UnityProject/Assets/Scripts/DTOs/PickupDTO.cs
@@ -1,14 +1,0 @@
-using System;
-
-[Serializable]
-public struct PickupDTO
-{
-    public string type;
-    public Location location;
-
-    public PickupType pickupType {
-        get {
-            return PickupTypeMethods.PickupTypeFromString(type);
-        }
-    }
-}
