--- UnityProject/Assets/Scripts/DTOs/PickupType.cs
@@ -1,7 +1,0 @@
-
-    public enum PickupType
-    {
-        Health,
-        Invulnerability,
-        DamageBoost
-    }
