--- UnityProject/Assets/Scripts/DTOs/PickupTypeMethods.cs
@@ -1,18 +1,0 @@
-
-    public static class PickupTypeMethods
-    {
-        public static PickupType PickupTypeFromString(string pickupType)
-        {
-            switch (pickupType.ToLower())
-            {
-                case "health":
-                    return PickupType.Health;
-                case "invulnerability":
-                    return PickupType.Invulnerability;
-                case "damage_boost":
-                    return PickupType.DamageBoost;
-                default:
-                    return PickupType.Health;
-            }
-        }
-    }
