--- UnityProject/Assets/Scripts/DTOs/PickupsDTO.cs
@@ -1,5 +1,0 @@
-using System;
-[Serializable]
-public struct PickupsDTO {
-    public PickupDTO[] pickups;
-}
