--- UnityProject/Assets/Scripts/DTOs/ScoreLocationDTO.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-[Serializable]
-public struct ScoreLocationDTO
-{
-    public Location location;
-}
