--- UnityProject/Assets/Scripts/DTOs/ScoreLocationsDTO.cs
@@ -1,7 +1,0 @@
-﻿using System;
-
-
-public struct ScoreLocationsDTO
-{
-    public ScoreLocationDTO[] scoreLocations;
-}
