+++ UnityProject/Assets/Scripts/MapFeatures/Score Locations/ScoreLocationDTO.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace MapFeatures.ScoreLocations
+{
+    [Serializable]
+    public struct ScoreLocationsDTO
+    {
+        public ScoreLocationDTO[] scoreLocations;
+    }
+
+    [Serializable]
+    public struct ScoreLocationDTO
+    {
+        public Location location;
+    }
+   
+}
