+++ UnityProject/Assets/Scripts/Shared DTOs/Location.cs
