+++ UnityProject/Assets/Scripts/Shared DTOs/ObstacleDTO.cs.meta
