+++ UnityProject/Assets/Scripts/Shared DTOs/ObstaclesDTO.cs.meta
