+++ UnityProject/Assets/Scripts/Shared DTOs/Orientation.cs.meta
