+++ UnityProject/Assets/Scripts/Shared DTOs/PickupDTO.cs.meta
