+++ UnityProject/Assets/Scripts/Shared DTOs/PickupType.cs.meta
