+++ UnityProject/Assets/Scripts/Shared DTOs/PickupTypeMethods.cs.meta
