+++ UnityProject/Assets/Scripts/Shared DTOs/PickupsDTO.cs.meta
