+++ UnityProject/Assets/Scripts/Shared DTOs/ScoreLocationDTO.cs.meta
