+++ UnityProject/Assets/Scripts/Shared DTOs/ScoreLocationsDTO.cs.meta
